TFBMTool-alt: switch to unicode. And drop Linux support.
Because opening files is hard. If I want unicode, Windows knows only
_wfopen/_tfopen (with wchar_t*/TCHAR* as parameter), and Linux knows
only fopen with a UTF-8 char* as parameter. I could rewrite a tchar.h
for linux, but that's too much work for my needs. I work on Windows, if
you want to use this tool on Linux, send a pull request.