Add th135arc-alt - a full rewrite of th135arc in C++, using classes.
The 1st goal is to finally have something that makes th135 support easy to maintain. I ended up dropping support in the old tool, because I would have needed to put if (is_th135) everywhere. With inheritance, I can do something better.
The 2nd goal is to have both Linux and Unicode support. If I want unicode support, the way to go on Windows is to use wchar_t. But on the other hand, on Linux, you're supposed to use UTF-8, and the wchar_t support is quite poor (it doesn't have _wfopen, which is mandatory for this tool). So let's write wrappers.
And the 3rd goal is to provide an API, instead of only a extract_the_stuff function and a plz_make_a_pak_file function. The Palette Editor needs to extract on the disk a whole gigabyte of data, only to get some sprites and palettes. And this is sad. With this, I'll be able to rewrite this part of the palette editor, and have it extract only the files it needs. And maybe some other projects could make use of that, the nut debugger for example?

Support for creating archives is planned, but not as high priority, because you can replace the files with thcrap - and that's what I do, so I don't need it myself. I would only do it for the sake of providing a complete tool. Which means I will probably have a lot of more important things to do before.