bmpfont_create: add an API to use as a library.
Because I had to copy-paste most of its source code in thcrap, and both
bmpfont_create and thcrap lost things with this.
I also changed the plugins API because to make it easier to call from a
library, and because I'll need to have some binary options
(--font-memory).
And about --font-memory, I imported here the things I hacked into thcrap
earlier: loading a font from an in-memory file, UTF-8, bold, and maybe I
or 2 more things.
I didn't test a lot, I'll test harder when I'll integrate it into thcrap