bmpfont: compute the chars count later.
There was a bug, when a character had to be redrawn on the next line,
where the chars count would not be updated accordingly, which caused the
metadatas to be totally unreadable. To fix this, I'll put the chars
count in a loop with less branches, because this seems less error-prone.