th145arc: fix a bug with /p and unicode strings.
"Unicode strings" here doesn't mean "strings with weird characters", but
only "strings encoded as UTF-16". And we convert the input path to
UTF-16. So I don't know why we didn't notice it was broken when we used
it back in th145 static patching... Maybe I made the unicode
improvements after the static patch?