th135arc-alt: replace UString with std::filesystem::path
C++ fucking delivered. We now have a standard class able to deal with paths in local encoding, in UTF-8 and in UTF-16. And std::ifstream/ofstream can take an unicode std::filesystem::path instead of a locale-dependant char*. And as a nice bonus, we also have standard functions to create directories. I can wipe almost all the platform-specific code I wrote and replace it with a few calls to these classes.